Move more V1 API stuff to V1 tree
* validator_filter_string
* DEPRECATED_FILTER_NAMES

Get rid of an unused Filter model attribute in the same go.
